requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
autopep8==2.1.0
certifi==2024.2.2
colorama==0.4.6
distro==1.9.0
Django==5.0.4
django-extensions==3.2.3
django-seed==0.3.1
django-silk==5.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.2.1
Faker==24.11.0
gprof2dot==2022.7.29
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
openai==1.23.5
pillow==10.3.0
psycopg2==2.9.9
pycodestyle==2.11.1
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.35.0
rpds-py==0.18.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tomli==2.0.1
toposort==1.10
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
